blue position 2 Updates!
Moved the value changer to a secondary button, wrote a function to
change the value and text on a passed pair of variable and label, and
got it to change to a fixed value.  Need to replicate on other defenses.

Still need to work out actually picking a defense and packing the
selections into the output string.